refactor file rw to utils
