+++ src/include/utils.hpp
@@ -1,0 +1,31 @@
+#pragma once
+#include <fstream>
+#include <vector>
+
+namespace utils {
+namespace infra {
+template <typename key_type>
+std::vector<key_type> read_txt(const char *filename) {
+    std::vector<key_type> data;
+    std::string line;
+    std::ifstream ifs(filename);
+    while (std::getline(ifs, line)) {
+        key_type key = std::stoul(line);
+        data.push_back(key);
+    }
+    return data;
+}
+
+template <typename key_type>
+std::vector<key_type> read_bin(const char *filename) {
+    std::ifstream inputFile(filename, std::ios::binary);
+    assert(inputFile.is_open());
+    inputFile.seekg(0, std::ios::end);
+    const std::streampos fileSize = inputFile.tellg();
+    inputFile.seekg(0, std::ios::beg);
+    std::vector<key_type> data(fileSize / sizeof(key_type));
+    inputFile.read(reinterpret_cast<char *>(data.data()), fileSize);
+    return data;
+}
+}  // namespace infra
+}  // namespace utils
